[Kashif]Added: BeanUtils and Swagger config

+++ src/main/java/com/cg/addressbook/service/IAddressBookService.java
@@ -9,11 +9,11 @@ public interface IAddressBookService {
 
 	List<AddressBookData> getAddressBookContactData();
 
-	AddressBookData getAddressBookContactDataById(int id);
+	AddressBookData getAddressBookContactDataById(Long id);
 
 	AddressBookData createAddressBookContactData(AddressBookDTO addressBookDTO);
 
-	AddressBookData updateAddressBookContactData(int id, AddressBookDTO addressBookDTO);
+	AddressBookData updateAddressBookContactData(Long id, AddressBookDTO addressBookDTO);
 
-	void deleteAddressBookContactData(int id);
+	void deleteAddressBookContactData(Long id);
 }
